pyproject.toml
[tool.poetry]
name = "magnet-monitor"
version = "0.1.0"
description = ""
authors = ["Antonello Pellecchia <antonello.pellecchia@cern.ch>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.3.4"
python-usbtmc = "^0.8"
pyusb = "^1.2.1"
dash = "^2.0.0"
dash-bootstrap-components = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
